[feature/COMMERCE-11] Koin framework and starting from it added in module "app".
