+++ app/build.gradle
@@ -50,6 +50,16 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.4.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
 
+    // Koin AndroidX Scope features
+    implementation "org.koin:koin-androidx-scope:$koin_version"
+    // Koin AndroidX ViewModel features
+    implementation "org.koin:koin-androidx-viewmodel:$koin_version"
+    // Koin AndroidX Fragment features
+    implementation "org.koin:koin-androidx-fragment:$koin_version"
+    implementation "org.koin:koin-android-architecture:0.8.2"
+    // Koin AndroidX Experimental features
+    //implementation "org.koin:koin-androidx-ext:$koin_version"
+
     kapt "android.arch.persistence.room:compiler:1.0.0"
 
     testImplementation 'junit:junit:4.12'
