--- login-user/src/main/java/com/manday/loginuser/CommerceApp.kt
@@ -1,20 +1,0 @@
-package com.manday.loginuser
-
-import android.app.Application
-import com.manday.loginuser.di.loginUserModule
-import org.koin.android.ext.koin.androidContext
-import org.koin.android.ext.koin.androidLogger
-import org.koin.core.context.startKoin
-
-class CommerceApp: Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        // Start Koin
-        startKoin{
-            androidLogger()
-            androidContext(this@CommerceApp)
-            modules(loginUserModule)
-        }
-    }
-}
