+++ login-user/src/main/java/com/manday/loginuser/domain/LoginRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.manday.loginuser.domain
 
 import com.manday.loginuser.data.LoginDataSource
 
-class LoginRepositoryImpl(var dataSource: LoginDataSource): LoginRepository {
+internal class LoginRepositoryImpl(var dataSource: LoginDataSource): LoginRepository {
 
     override fun login(user: String, pass: String) {
         dataSource.login(user, pass)
