+++ login-user/src/main/java/com/manday/loginuser/domain/interactors/LoginEmployeeInteractor.kt
@@ -2,7 +2,7 @@ package com.sdos.login.domain
 
 import com.manday.loginuser.domain.LoginRepository
 
-class LoginEmployeeInteractor(var repository: LoginRepository): (String, String) -> Unit {
+internal class LoginEmployeeInteractor(var repository: LoginRepository): (String, String) -> Unit {
 
     override fun invoke(user: String, pass: String) {
         return repository.login(user, pass)
